Rename face detector classes and assets for clarity
Renamed Java classes and related assets to use more descriptive names for face detection (CaffeFaceDetector, CaffeFaceDetectorWebcam, HaarCascadeFaceDetector). Updated image and resource paths to reflect new structure. Removed unused image files to clean up the project.
